fix(evil): fix autoloading order.
When a function is called even in a `eval-after-load` form, then those
autoloaded function is being loaded.

That is, I define a function in my autoloads file, and the
`eval-after-load` form in my init file calls this function, then the
autoloaded file is loaded. And at that tiime evil isn't
required. Which causes the loading error.

Fix: require evil directly before the `eval-after-load` form.