fix(copilot): use `M->` instead of `M-]` for next suggestion.
In tty emacs, bind `M-]` and `M-[` to a command may cause weird
behavior such that when scrolling the screen with touchpad, the buffer
will be filled with weird control sequence characters.  This is
because the key sequence of special keys like `<f5>` to `<f12>` are
prefixed with `ESC [` which is exactly the key sequence of `M-[`. Thus
bind `M->` and `M-<` instead of `M-]` and `M-]`. Vanilla emacs user
may highly rely on `M-<` and `M->` to nagivate between the beginning
and end of the buffer. But I use evil, so I am a `gg` and `G` man.

See discussion here:
https://emacs.stackexchange.com/questions/68703/m-causes-emacs-to-print-weird-possibly-escape-sequences.