feat(code-cells): add `code-cells` for editing jupyter notebook.
`code-cells` is a bridge between the plain text format file and the
`.ipynb` file. It allows you to edit the plain text file in emacs and
then convert it to the `.ipynb` file. The `.ipynb` file can be opened
in jupyter notebook. A plain text file and a `.ipynb` file is synced
by `code-cells` automatically.

To use this feature, you need to install the command line tool
`jupytext`.