Revert "feat!(mu4e): use built-in thread folding capability."
There is a bug in the official thread folding implementation, such as
the entire thread (including the first message) folding when opening the
first message from a folded thread.

Although my custom implementation lacks a deep understanding of mu4e
internals and relies on some hacks (such as pledging to a timer in the
hope of a magical resolution), it works, anyway.

This reverts commit 03bf041289dff89c4ba0daf84cef6cdfaeac9541.