config(tab-bar): update `tab-bar-new-tab-choice`
Previously, this option was set to `"*scratch*"`, which would create a
buffer associated with a file (named `*scratch*`) if that buffer did
not exist. This behavior was not ideal, as one might accidentally save
that file, resulting in directory pollution. Set this option to a
function that creates a buffer and avoid associating it with a file.