feat(minuet): prompt template and few shots can be a symbol or function.
When the prompt template or few shots is a symbol, evaluate it at the
run time to get its value, which can be a string or a list (for few
shots).

When the prompt template or few shots is a function, call it at the
run time to get its value, which can be a string or a list (for few
shots).

This allows the user to customize the prompt template and few shots
based on the current context. For example, bind
`minuet-default-fewshots` and `minuet-default-system-template` to a
variable with different local value depending on the major mode.